Create basic Reconcile methods for WatcherAPI
This patch is basic recond methods to the watcherapi
controller, as well as some initial input validation.

The change adds the required CR spec and status, and creates the initial
structure for initialization and deletion. In addition it some some
input validation by accessing the osp secret and the database that
should be created by the watcher controller.

Finally, it is also adding some initial structure test for functional envtest
testing in WatcherAPI.

Related: OSPRH-11483