Create database schema for Watcher database
This commit is implementing db schema creation (dbsync) via Job
following the same pattern as other operators.

It introduces some spec and status changes required and initial
configuration with only the required configuration required to run the
dbsync (just the database connection and transport which will be
required later to pass it to the subCRs).

This patch is also adding a new functional scenario test to test the
Watcher controller when using non-default values.

Finally, this patch is also skipping two new error types in the
crd-schema-checker as it complains about perserveJobs for being a
boolean type and Hash to be a map. However we want to keep them with
those types for consistency with the rest of operators.