Add a github action to sync content from main to fr1 branch
In order to align with the resto of operators we need to have a 18.0-fr1
branch. However, we do not want to manually cherry-pick every commit but
to get any new commit into the branch until we have a fully functional
repo.

This patch is adding a job which will sync content from main to the branch
18.0-fr1 after any merge into the main branch.