Enable watcher_tempest_plugin tests in validation jobs
We have decision-engine and applier integrated in the Watcher controller
so we can increase tempest coverage. I am excluding some tests from the
watcher plugin:

- client_functional tests require watcherclient installed in the tempest
  container.
- Some strategies execute tests are failing with an error while accessing
  resource traits for compute nodes. Until we can manage this, we can
  exclude these tests.

Also this patch is setting some tempest parameters aligned with upstream
ones and forcing to use the watcher-tempest-pluging from upstream master
branch source instead of rpm.

Finally, this patch is also creating a new Watcher CR file appropiate
for CI usage and modifies Makefile and the watcher deployment playbook
to customize the path of the CR when using `make watcher_deploy`.